Compliance with mapping_table defined in release 2462
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2468 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/module/internal_component/vci_vcache_wrapper/caba/source/include/vci_vcache_wrapper.h
@@ -200,7 +200,7 @@ public:
 private:
 
     // STRUCTURAL PARAMETERS
-    soclib::common::AddressDecodingTable<uint32_t, bool>    	m_cacheability_table;
+    soclib::common::AddressDecodingTable<uint64_t, bool>    	m_cacheability_table;
     const vci_srcid_t                                       	m_srcid;
 
     const size_t  						m_tlb_ways;
